+++ volmdlr/surfaces.py
@@ -4948,148 +4948,73 @@ class BSplineSurface3D(Surface3D):
         """
         Converts the primitive from 3D spatial coordinates to its equivalent 2D primitive in the parametric space.
         """
-        # TODO: enhance this, it is a non exact method!
-        # TODO: bsplinecurve can be periodic but not around the bsplinesurface
-        flag = False
-        if not bspline_curve3d.points[0].is_close(bspline_curve3d.points[-1]):
-            bsc_linesegment = edges.LineSegment3D(bspline_curve3d.points[0],
-                                                  bspline_curve3d.points[-1])
-            flag = True
-            for point in bspline_curve3d.points:
-                if not bsc_linesegment.point_belongs(point):
-                    flag = False
-                    break
+        lth = bspline_curve3d.length()
 
-        if self.x_periodicity and not self.y_periodicity \
-                and bspline_curve3d.periodic:
-            point1 = self.point3d_to_2d(bspline_curve3d.points[0])
-            p1_sup = self.point3d_to_2d(bspline_curve3d.points[0])
-            new_x = point1.x - p1_sup.x + self.x_periodicity
-            new_x = new_x if 0 <= new_x else 0
-            reverse = False
-            if new_x < 0:
-                new_x = 0
-            elif math.isclose(new_x, self.x_periodicity, abs_tol=1e-5):
-                new_x = 0
-                reverse = True
+        if lth <= 1e-6:
+            print('BSplineCurve3D skipped because it is too small')
+            return []
 
-            linesegments = [
-                edges.LineSegment2D(
-                    volmdlr.Point2D(new_x, point1.y),
-                    volmdlr.Point2D(self.x_periodicity, point1.y))]
-            if reverse:
-                linesegments[0] = linesegments[0].reverse()
+        n = min(len(bspline_curve3d.control_points), 20)  # Limit points to avoid non-convergence
+        points3d = bspline_curve3d.discretization_points(number_points=n)
+        points = [self.point3d_to_2d(p) for p in points3d]
 
-        elif self.y_periodicity and not self.x_periodicity \
-                and bspline_curve3d.periodic:
-            point1 = self.point3d_to_2d(bspline_curve3d.points[0])
-            p1_sup = self.point3d_to_2d(bspline_curve3d.points[0])
-            new_y = point1.y - p1_sup.y + self.y_periodicity
-            new_y = new_y if 0 <= new_y else 0
-            reverse = False
-            if new_y < 0:
-                new_y = 0
-            elif math.isclose(new_y, self.y_periodicity, abs_tol=1e-5):
-                new_y = 0
-                reverse = True
+        if self.u_closed() or self.v_closed():
+            points = self.check_start_end_parametric_points(bspline_curve3d, points, points3d)
 
-            linesegments = [
-                edges.LineSegment2D(
-                    volmdlr.Point2D(point1.x, new_y),
-                    volmdlr.Point2D(point1.x, self.y_periodicity))]
-            if reverse:
-                linesegments[0] = linesegments[0].reverse()
+        if self.x_periodicity:
+            points = self._repair_periodic_boundary_points(bspline_curve3d, points, 'x')
+            if bspline_curve3d.periodic:
+                points = self._handle_periodic_curve(bspline_curve3d, points, 'x')
 
-        elif self.x_periodicity and self.y_periodicity \
-                and bspline_curve3d.periodic:
-            raise NotImplementedError
+        if self.y_periodicity:
+            points = self._repair_periodic_boundary_points(bspline_curve3d, points, 'y')
+            if bspline_curve3d.periodic:
+                points = self._handle_periodic_curve(bspline_curve3d, points, 'y')
 
-        if flag:
-            x_perio = self.x_periodicity if self.x_periodicity is not None \
-                else 1.
-            y_perio = self.y_periodicity if self.y_periodicity is not None \
-                else 1.
+        if self._is_line_segment(points):
+            return [edges.LineSegment2D(points[0], points[-1])]
 
-            point1 = self.point3d_to_2d(bspline_curve3d.points[0])
-            point2 = self.point3d_to_2d(bspline_curve3d.points[-1])
+        return [edges.BSplineCurve2D.from_points_interpolation(points=points, degree=bspline_curve3d.degree)]
 
-            if point1.is_close(point2):
-                print('BSplineCruve3D skipped because it is too small')
-                linesegments = None
+    @staticmethod
+    def _handle_periodic_curve(bspline_curve3d, points, axis):
+        u_min, u_max = bspline_curve3d.curve.domain
+        start_param = points[0].x if axis == 'x' else points[0].y
+        param_after_start = points[1].x if axis == 'x' else points[1].y
+        should_be_umax = (u_max - param_after_start) < (param_after_start - u_min)
+        if math.isclose(start_param, u_min, abs_tol=1e-6):
+            if axis == 'x':
+                if should_be_umax:
+                    points[0] = volmdlr.Point2D(u_max, points[0].y)
+                else:
+                    points[-1] = volmdlr.Point2D(u_max, points[-1].y)
             else:
-                p1_sup = self.point3d_to_2d(bspline_curve3d.points[0])
-                p2_sup = self.point3d_to_2d(bspline_curve3d.points[-1])
-                if self.x_periodicity and point1.point_distance(p1_sup) > 1e-5:
-                    point1.x -= p1_sup.x - x_perio
-                    point2.x -= p2_sup.x - x_perio
-                if self.y_periodicity and point1.point_distance(p1_sup) > 1e-5:
-                    point1.y -= p1_sup.y - y_perio
-                    point2.y -= p2_sup.y - y_perio
-                linesegments = [edges.LineSegment2D(point1, point2)]
-            # How to check if end of surface overlaps start or the opposite ?
-        else:
-            lth = bspline_curve3d.length()
-            if lth > 1e-5:
-                n = min(len(bspline_curve3d.control_points), 20) # limit points to avoid non convergence
-                points3d = bspline_curve3d.discretization_points(number_points=n)
-                points = [self.point3d_to_2d(p) for p in points3d]
-                if self.u_closed() or self.v_closed():
-                    points = self.check_start_end_parametric_points(bspline_curve3d, points, points3d)
-
-                if self.x_periodicity:
-                    points = self._repair_periodic_boundary_points(bspline_curve3d, points, 'x')
-                    if bspline_curve3d.periodic and points[0].is_close(points[-1]):
-                        u_min, u_max = bspline_curve3d.curve.domain
-                        if math.isclose(points[0].x, u_min, abs_tol=1e-6):
-                            should_be_umax = (u_max - points[1].x) < (points[1].x - u_min)
-                            if should_be_umax:
-                                points[0] = volmdlr.Point2D(u_max, points[0].y)
-                            else:
-                                points[-1] = volmdlr.Point2D(u_max, points[-1].y)
-                        elif math.isclose(points[0].x, u_max, abs_tol=1e-6):
-                            should_be_umin = (u_max - points[1].x) > (points[1].x - u_min)
-                            if should_be_umin:
-                                points[0] = volmdlr.Point2D(u_min, points[0].y)
-                            else:
-                                points[-1] = volmdlr.Point2D(u_min, points[-1].y)
-                if self.y_periodicity:
-                    points = self._repair_periodic_boundary_points(bspline_curve3d, points, 'y')
-                    if bspline_curve3d.periodic and points[0].is_close(points[-1]):
-                        u_min, u_max = bspline_curve3d.curve.domain
-                        if math.isclose(points[0].y, u_min, abs_tol=1e-6):
-                            should_be_umax = (u_max - points[1].y) < (points[1].y - u_min)
-                            if should_be_umax:
-                                points[0] = volmdlr.Point2D(points[0].x, u_max)
-                            else:
-                                points[-1] = volmdlr.Point2D(points[-1].x, u_max)
-                        elif math.isclose(points[0].y, u_max, abs_tol=1e-6):
-                            should_be_umin = (u_max - points[1].y) > (points[1].y - u_min)
-                            if should_be_umin:
-                                points[0] = volmdlr.Point2D(points[0].x, u_min)
-                            else:
-                                points[-1] = volmdlr.Point2D(points[-1].x, u_min)
-
-                if not points[0].is_close(points[-1]) and not bspline_curve3d.periodic:
-                    linesegment = edges.LineSegment2D(points[0], points[-1])
-                    flag_line = True
-                    for point in points:
-                        if not linesegment.point_belongs(point, abs_tol=1e-4):
-                            flag_line = False
-                            break
-                    if flag_line:
-                        return [linesegment]
-
-                return [edges.BSplineCurve2D.from_points_interpolation(points=points, degree=bspline_curve3d.degree)]
-
-            if 1e-6 < lth <= 1e-5:
-                linesegments = [edges.LineSegment2D(
-                    self.point3d_to_2d(bspline_curve3d.start),
-                    self.point3d_to_2d(bspline_curve3d.end))]
+                if should_be_umax:
+                    points[0] = volmdlr.Point2D(points[0].x, u_max)
+                else:
+                    points[-1] = volmdlr.Point2D(points[-1].x, u_max)
+        elif math.isclose(start_param, u_max, abs_tol=1e-6):
+            should_be_umin = (u_max - param_after_start) > (param_after_start - u_min)
+            if axis == 'x':
+                if should_be_umin:
+                    points[0] = volmdlr.Point2D(u_min, points[0].y)
+                else:
+                    points[-1] = volmdlr.Point2D(u_min, points[-1].y)
             else:
-                print('BSplineCruve3D skipped because it is too small')
-                linesegments = None
+                if should_be_umin:
+                    points[0] = volmdlr.Point2D(points[0].x, u_min)
+                else:
+                    points[-1] = volmdlr.Point2D(points[-1].x, u_min)
+        return points
 
-        return linesegments
+    @staticmethod
+    def _is_line_segment(points):
+        linesegment = edges.LineSegment2D(points[0], points[-1])
+        for point in points:
+            if not linesegment.point_belongs(point, abs_tol=1e-4):
+                return False
+        return True
+
 
     def bsplinecurve2d_to_3d(self, bspline_curve2d):
         """